Configure Vim, tmux and Ruby
The ideia is to keep all my configs, in the last months I got stuck on
this and had a lot of unstaged files to commit because I stopped in the
middle of a big change. The initial ideia was to separate this job into
several commits, but this shows (again) that this (larg commits) is not
a good ideia, it always take a big price.

This commits installs chruby, vim, tmux and other basic softwares as ack
and tree.

I promise will be more smart on next time (this is not the first time I
am doing this promise AHuAHuHA).